requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4
pypdf2==3.0.1
python-dotenv==1.0.0
tiktoken==0.5.1
